Currency-Fragments:  - Added extentions to build.gradle file for directly handling for UI component as global field  - Added CurrencyFragment.kt and fragment_currency.xml files for BottomNavigationView
